# perf(object-hash): faster object access by avoid string concat