Inspired by https://github.com/puleos/object-hash/pull/122

### perf: faster object access by avoid string concat

V8 from what I could understand can predict when the same pattern is happening, but when we insert some randomness, the performance of string concatening is a little bit worse.

```
'oldObjectAccess x 1,141,640,736 ops/sec ±0.53% (97 runs sampled)',
'newObjectAccess x 1,146,221,890 ops/sec ±0.33% (97 runs sampled)',
'random: oldObjectAccess x 9,099,441 ops/sec ±1.14% (92 runs sampled)',
'random: newObjectAccess x 23,251,465 ops/sec ±1.63% (88 runs sampled)'
```

> Benchmark: [bench-object-access.js](https://gist.github.com/H4ad/0c8b3816ccf3de1506095f7257f62816#file-bench-object-access-js)